hands_on_session/airflow/requirements.txt
sqlalchemy
pandas
click==8.1.7
psycopg2-binary
MDAnalysis==2.8.0
paramiko==3.4.0
apache-airflow-providers-postgres
h5py==3.12.1
# For mock_amber utility
numpy>=1.22
parmed>=4.2.2
--index-url https://opencode.it4i.eu/api/v4/projects/107/packages/pypi/simple
py4lexis==3.2.2